+++ src/blocks/store/s3.clj
@@ -29,7 +29,10 @@
       PutObjectRequest
       PutObjectResult
       S3Object
-      S3ObjectSummary)))
+      S3ObjectSummary)
+    (java.io
+      FilterInputStream
+      InputStream)))
 
 
 ;; ## S3 Utilities
@@ -78,6 +81,16 @@
     client))
 
 
+(def ^:private sse-algorithms
+  {:aes-256 ObjectMetadata/AES_256_SERVER_SIDE_ENCRYPTION})
+
+
+(defn- select-sse-algorithm
+  "Return corresponding SSE algorithm string constant or throw if not supported."
+  [algorithm]
+  (or (get sse-algorithms algorithm)
+      (throw (ex-info (format "Unsupported SSE algorithm '%s'" algorithm)
+                      {:supported (keys sse-algorithms) :given algorithm}))))
 
 ;; ## S3 Key Translation
 
@@ -127,6 +140,27 @@
    :s3/metadata (into {} (.getRawMetadata metadata))})
 
 
+(defn- auto-draining-stream
+  "Wraps an `InputStream` in a proxy which will automatically drain the
+  underlying stream when it is closed."
+  [^InputStream stream]
+  (proxy [FilterInputStream] [stream]
+    (close
+      []
+      ; TODO: be smarter about this; for large remaining payloads it may be
+      ; faster to abort and re-establish the connection than to drain the rest
+      ; of the object.
+      (let [start (System/nanoTime)]
+        (loop [drained 0]
+          (if (pos? (.read stream))
+            (recur (inc drained))
+            (when (pos? drained)
+              (log/debugf "Drained %d bytes in %.2f ms while closing S3 input stream"
+                          drained
+                          (/ (- (System/nanoTime) start) 1e6))))))
+      (.close stream))))
+
+
 (defn- object->block
   "Creates a lazy block to read from the given S3 object."
   [^AmazonS3 client ^String bucket prefix stats]
@@ -137,13 +171,17 @@
         (fn object-reader
           ([]
            (log/debugf "Opening object %s" (s3-uri bucket object-key))
-           (.getObjectContent (.getObject client bucket object-key)))
+           (->> (.getObject client bucket object-key)
+                (.getObjectContent)
+                (auto-draining-stream)))
           ([^long start ^long end]
            (log/debugf "Opening object %s byte range [%d, %d)"
                        (s3-uri bucket object-key) start end)
-           (let [request (doto (GetObjectRequest. bucket object-key)
-                           (.setRange start (dec end)))]
-             (.getObjectContent (.getObject client request)))))))
+           (->> (doto (GetObjectRequest. bucket object-key)
+                  (.setRange start (dec end)))
+                (.getObject client)
+                (.getObjectContent)
+                (auto-draining-stream))))))
     (dissoc stats :id :size)))
 
 
@@ -176,7 +214,9 @@
 (defrecord S3BlockStore
   [^AmazonS3 client
    ^String bucket
-   ^String prefix]
+   ^String prefix
+   sse
+   alter-put-metadata]
 
   store/BlockStore
 
@@ -221,9 +261,12 @@
         (object->block client bucket prefix stats)
         ; Otherwise, upload block to S3.
         (let [object-key (id->key prefix (:id block))
-              ; TODO: support function which allows you to mutate the ObjectMetadata before it is sent in the putObject call?
               metadata (doto (ObjectMetadata.)
                          (.setContentLength (:size block)))]
+          (when sse
+            (.setSSEAlgorithm metadata (select-sse-algorithm sse)))
+          (when alter-put-metadata
+            (alter-put-metadata this metadata))
           (log/debugf "PutObject %s to %s" block (s3-uri bucket object-key))
           (let [result (with-open [content (block/open block)]
                          (.putObject client bucket object-key content metadata))
@@ -291,7 +334,12 @@
   - `:credentials` a map with `:access-key` and `:secret-key` entries providing
     explicit AWS credentials.
   - `:region` a keyword or string designating the region the bucket is in.
-  - `:prefix` a string prefix to store the blocks under."
+  - `:prefix` a string prefix to store the blocks under.
+  - `:sse` a keyword algorithm selection to set Server Side Encryption
+    on block PUT. No `:sse` present will not set this flag.
+  - `:alter-put-metdata` a 2-arity function that operates on the block store
+    record and this metadata. This function is called before a put operation and
+    the return value is discarded. Content-Length metadata is already specified."
   [bucket & {:as opts}]
   (when (or (not (string? bucket))
             (empty? (str/trim bucket)))
@@ -313,6 +361,10 @@
       (:host uri)
       :prefix (:path uri)
       :region (keyword (get-in uri [:query :region]))
+      :sse (when-let [algorithm (keyword (get-in uri [:query :sse]))]
+             ;; check if supported, but return keyword
+             (select-sse-algorithm algorithm)
+             algorithm)
       :credentials (when-let [creds (:user-info uri)]
                      {:access-key (:id creds)
                       :secret-key (:secret creds)}))))