+++ CHANGELOG.md
@@ -9,6 +9,25 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ...
 
+## [0.4.0] - 2018-03-22
+
+### Changed
+- Upgrade Clojure to 1.9.0
+- Upgrade Blocks to 1.1.0
+- Upgrade S3 SDK to 1.11.271
+
+### Added
+- S3 Block store supports basic server-side encryption (SSE) by setting the
+  `:sse` attribute to `:aes-256`.
+  [#4](https://github.com/greglook/blocks-s3/pull/4)
+- An arbitrary `alter-put-metadata` function can be provided to mutate block
+  object metadata before it is written.
+
+### Fixed
+- S3 block content streams will now automatically drain the remaining bytes in
+  the stream when they are closed.
+  [#2](https://github.com/greglook/blocks-s3/issues/2)
+
 ## [0.3.3] - 2017-04-28
 
 ### Changed
@@ -51,7 +70,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial library release.
 
-[Unreleased]: https://github.com/greglook/blocks-s3/compare/0.3.3...HEAD
+[Unreleased]: https://github.com/greglook/blocks-s3/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/greglook/blocks-s3/compare/0.3.3...0.4.0
 [0.3.3]: https://github.com/greglook/blocks-s3/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/greglook/blocks-s3/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/greglook/blocks-s3/compare/0.3.0...0.3.1